+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/ManualCleanupPage.kt
@@ -2,8 +2,8 @@ package au.com.shiftyjelly.pocketcasts.settings
 
 import android.content.Context
 import android.widget.Toast
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.selection.toggleable
@@ -25,7 +25,7 @@ import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import au.com.shiftyjelly.pocketcasts.compose.AppTheme
+import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.bars.NavigationButton
 import au.com.shiftyjelly.pocketcasts.compose.bars.ThemedTopAppBar
 import au.com.shiftyjelly.pocketcasts.compose.buttons.RowButton
@@ -89,8 +89,8 @@ private fun ManageDownloadsView(
     val deleteButtonColor = MaterialTheme.theme.colors.support05
     Column(
         modifier = modifier
-            .background(MaterialTheme.theme.colors.primaryUi01)
             .padding(top = 8.dp)
+            .fillMaxHeight()
             .verticalScroll(rememberScrollState())
     ) {
         state.diskSpaceViews.forEach { DiskSpaceSizeRow(it, onDiskSpaceCheckedChanged) }
@@ -185,7 +185,7 @@ private fun ManualCleanupPageSmallPreview() {
 private fun ManualCleanupPagePreview(
     themeType: Theme.ThemeType = Theme.ThemeType.LIGHT
 ) {
-    AppTheme(themeType) {
+    AppThemeWithBackground(themeType) {
         ManageDownloadsView(
             state = ManualCleanupViewModel.State(),
             includeStarredSwitchState = false,