+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/opml/OpmlImportTask.kt
@@ -29,8 +29,13 @@ import kotlinx.coroutines.flow.mapNotNull
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.toList
 import org.xml.sax.Attributes
+import org.xml.sax.InputSource
+import org.xml.sax.SAXException
 import org.xml.sax.helpers.DefaultHandler
 import java.io.InputStream
+import java.io.StringReader
+import java.util.Scanner
+import java.util.regex.Pattern
 import javax.xml.parsers.SAXParserFactory
 import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
@@ -62,6 +67,64 @@ class OpmlImportTask @AssistedInject constructor(
 
             Toast.makeText(context, context.getString(LR.string.settings_import_opml_toast), Toast.LENGTH_LONG).show()
         }
+
+        /**
+         * Extract the feed urls from an OPML file using a SAX Parser.
+         */
+        fun readOpmlUrlsSax(inputStream: InputStream): List<String> {
+            val parser = SAXParserFactory.newInstance().newSAXParser()
+            val urls = mutableListOf<String>()
+            parser.parse(
+                inputStream,
+                object : DefaultHandler() {
+                    override fun startElement(
+                        uri: String?,
+                        localName: String?,
+                        qName: String?,
+                        attributes: Attributes?
+                    ) {
+                        if (localName.equals("outline", ignoreCase = true)) {
+                            val url = attributes?.getValue("xmlUrl")
+                            if (!url.isNullOrBlank()) {
+                                urls.add(url)
+                            }
+                        }
+                    }
+                }
+            )
+            return urls
+        }
+
+        /**
+         * Extract the feed urls from an OPML file using a Regex.
+         */
+        fun readOpmlUrlsRegex(inputStream: InputStream): List<String> {
+            val urls = mutableListOf<String>()
+            val scanner = Scanner(inputStream, "UTF-8")
+            val pattern = Pattern.compile("xmlUrl=\"([^\"]+)\"")
+            while (scanner.findWithinHorizon(pattern, 0) != null) {
+                val result = scanner.match()
+                val url = encodeXmlString(result.group(1))
+                urls.add(url)
+            }
+            return urls
+        }
+
+        private fun encodeXmlString(url: String): String {
+            val parser = SAXParserFactory.newInstance().newSAXParser()
+            val decodedUrl = StringBuilder()
+            parser.parse(
+                InputSource().apply {
+                    characterStream = StringReader("<root>$url</root>")
+                },
+                object : DefaultHandler() {
+                    override fun characters(charArray: CharArray, start: Int, length: Int) {
+                        decodedUrl.append(String(charArray, start, length))
+                    }
+                }
+            )
+            return decodedUrl.toString()
+        }
     }
 
     private val notificationManager = context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
@@ -69,9 +132,7 @@ class OpmlImportTask @AssistedInject constructor(
     override suspend fun doWork(): Result {
         try {
             val uri = Uri.parse(inputData.getString(INPUT_URI)) ?: return Result.failure()
-            applicationContext.contentResolver.openInputStream(uri)?.use { inputStream ->
-                processFile(inputStream)
-            }
+            processFile(uri)
             return Result.success()
         } catch (t: Throwable) {
             LogBuffer.e(LogBuffer.TAG_BACKGROUND_TASKS, t, "OPML import failed.")
@@ -79,8 +140,19 @@ class OpmlImportTask @AssistedInject constructor(
         }
     }
 
-    private suspend fun processFile(inputStream: InputStream) {
-        val urls = readOpmlFeedUrls(inputStream)
+    private suspend fun processFile(uri: Uri) {
+        var urls = emptyList<String>()
+
+        val resolver = applicationContext.contentResolver
+        try {
+            resolver.openInputStream(uri)?.use { inputStream ->
+                urls = readOpmlUrlsSax(inputStream)
+            }
+        } catch (e: SAXException) {
+            resolver.openInputStream(uri)?.use { inputStream ->
+                urls = readOpmlUrlsRegex(inputStream)
+            }
+        }
 
         val podcastCount = urls.size
         val initialDatabaseCount = podcastManager.countPodcasts()
@@ -119,28 +191,6 @@ class OpmlImportTask @AssistedInject constructor(
             updateNotification(initialDatabaseCount, podcastCount)
             delay(1000)
         }
-    }
-
-    /**
-     * Extract the feed urls from an OPML file.
-     */
-    private fun readOpmlFeedUrls(inputStream: InputStream): List<String> {
-        val parser = SAXParserFactory.newInstance().newSAXParser()
-        val urls = mutableListOf<String>()
-        parser.parse(
-            inputStream,
-            object : DefaultHandler() {
-                override fun startElement(uri: String?, localName: String?, qName: String?, attributes: Attributes?) {
-                    if (localName.equals("outline", ignoreCase = true)) {
-                        val url = attributes?.getValue("xmlUrl")
-                        if (!url.isNullOrBlank()) {
-                            urls.add(url)
-                        }
-                    }
-                }
-            }
-        )
-        return urls
     }
 
     /**